+++ src/components/AllPagesPanel/index.ts
@@ -1,3 +1,0 @@
-export {AllPagesPanel} from './AllPagesPanel';
-export {getAllPagesMenuItem, ALL_PAGES_ID} from './constants';
-export {useVisibleMenuItems} from './useVisibleMenuItems';