+++ src/components/AsideHeader/components/AllPagesPanel/index.ts
@@ -1,0 +1,3 @@
+export {AllPagesPanel} from './AllPagesPanel';
+export {getAllPagesMenuItem} from './constants';
+export {useVisibleMenuItems} from './useVisibleMenuItems';