+++ src/components/AsideHeader/components/CompositeBar/MultipleTooltip/index.ts
@@ -1,0 +1,2 @@
+export {MultipleTooltip} from './MultipleTooltip';
+export {MultipleTooltipContext, MultipleTooltipProvider} from './MultipleTooltipContext';