+++ src/components/Settings/__stories__/SettingsMobileDemo.tsx
@@ -163,6 +163,13 @@ export const SettingsMobileComponent = React.memo(
                                     </Button>
                                 </Settings.Item>
                             )}
+                            <Settings.Item
+                                title="Full width"
+                                id="full-width"
+                                renderTitleComponent={() => null}
+                            >
+                                Place to display full width content
+                            </Settings.Item>
                         </Settings.Section>
                     </Settings.Page>
                     <Settings.Page id="appearance" title="Appearance">