+++ src/components/types.ts
@@ -2,8 +2,6 @@ import React, {HTMLAttributeAnchorTarget} from 'react';
 
 import {AlertProps, IconProps, QAProps} from '@gravity-ui/uikit';
 
-import {ItemProps} from './CompositeBar/Item/Item';
-
 export type MenuItemType = 'regular' | 'action' | 'divider';
 
 export type OpenModalSubscriber = (open: boolean) => void;
@@ -20,7 +18,7 @@ export interface MenuItem extends QAProps {
     icon?: IconProps['data'];
     iconSize?: number | string;
     iconQa?: string;
-    link?: string;
+    href?: string;
     current?: boolean;
     pinned?: boolean;
     onItemClick?: (
@@ -29,7 +27,6 @@ export interface MenuItem extends QAProps {
         event: React.MouseEvent<HTMLElement, MouseEvent>,
     ) => void;
     onItemClickCapture?: (event: React.SyntheticEvent) => void;
-    onCollapseItemClick?: () => void;
     itemWrapper?: (
         p: MakeItemParams,
         makeItem: (p: MakeItemParams) => React.ReactNode,
@@ -59,8 +56,6 @@ export interface MenuItem extends QAProps {
     className?: string;
 }
 
-export type SubheaderMenuItem = Omit<ItemProps, 'onItemClick' | 'onItemClickCapture'>;
-
 export interface LogoProps {
     text: (() => React.ReactNode) | string;
     className?: string;
@@ -77,7 +72,7 @@ export interface LogoProps {
     'aria-labelledby'?: string;
 }
 
-export interface TopAlertProps {
+interface TopAlertBaseProps {
     align?: AlertProps['align'];
     message: AlertProps['message'];
     title?: AlertProps['title'];
@@ -89,4 +84,11 @@ export interface TopAlertProps {
     centered?: boolean;
     dense?: boolean;
     onCloseTopAlert?: () => void;
+    render?: never;
 }
+
+export type TopAlertProps =
+    | TopAlertBaseProps
+    | {
+          render: (params: {handleClose: () => void}) => React.ReactElement;
+      };